Added ims apk __>>package as a system privileged app
The mk file installs the apk as a privilege app. Radio came back after this was added as privileged app.
Place the folder with mk file in packages/apps/ims.apk
then add the apk to the app list in the handheld_system.mk file like this ims \ --> keep it in alphabetical order.
There is also two permission files that you need or the device will not boot. Use the permission files provided. The two permission files go in system/etc/permissions